Added goArray
Added comments to goSwitch. Started and finished goArray.
goArray:
Here we create an array a that will hold exactly 5 ints. The type of
elements and length are both part of the array’s type. By default an
array is zero-valued, which for ints means 0s.

Note that arrays appear in the form [v1 v2 v3 ...] when printed with
fmt.Println.